Add error popups and safe_fetch() wrapper function
This commit adds a globally available ErrorModal that only appears when
the `error` key in the Redux store is set.
`error` is emptied when the ErrorModal is dismissed (by either clicking
the X button, or by clicking off of the popup, or by pressing Escape).

To abstract the implementation of store writes and request exception
handling, the `safe_fetch()` function, which wraps around the native
`fetch()` function, handles basic sanity checking, and returns the
response JSON if there is no error. In the case of an error, it will
throw any exceptions from `fetch()`, or the `message` field from a
response with a non-2XX status.

Known errata: The use of the Redux store triggers repeated fetching from
the draft API endpoints, as there is a subscription on the draft pages
doing so. My current proposal is to rate-limit this request, and cache
the result on the client-side for a minimum period of time.